refactor(ICache): move ICacheImp into separate file
